feat(auction): add exception for auctions with lots
* Introduced AuctionHasLotsException to handle cases
  where an auction cannot be deleted if it has associated lots.
* Updated delete method in AuctionService to check for
  existing lots before deletion and throw the new exception.

+++ src/main/java/br/com/leje/lancecerto/auction/AuctionService.java
@@ -3,6 +3,8 @@ package br.com.leje.lancecerto.auction;
 import br.com.leje.lancecerto.auction.dto.AuctionRequest;
 import br.com.leje.lancecerto.auction.dto.AuctionResponse;
 import br.com.leje.lancecerto.auction.dto.AuctionUpdateStatusRequest;
+import br.com.leje.lancecerto.lot.LotRepository;
+import br.com.leje.lancecerto.shared.exception.AuctionHasLotsException;
 import br.com.leje.lancecerto.shared.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,6 +19,7 @@ public class AuctionService {
 
     private final AuctionMapper mapper;
     private final AuctionRepository repository;
+    private final LotRepository lotRepository;
 
     public AuctionResponse create(AuctionRequest request) {
         Auction auction = mapper.toEntity(request);
@@ -52,7 +55,11 @@ public class AuctionService {
     @Transactional
     public void delete(UUID id) {
         Auction auction = findByIdOrThrow(id);
-        // TODO Fase Lot: bloquear 409 se houver lotes vinculados
+        var hasLots = lotRepository.existsByAuctionId(auction.getId());
+
+        if (hasLots) {
+            throw new AuctionHasLotsException();
+        }
 
         repository.delete(auction);
     }
